requirements.txt
python-dateutil==2.8.2
six==1.16.0
psycopg2-binary==2.9.9
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1 